[BREAKING] Use doubles for lat/lon
Because single-precision floating point only gives ~1m
precision for most latitude/longitudes, I am switching to doubles.
Because the Airmar actually supplies lat/lon in single-precision
floats and because GPS isn't any more accurate than that (for us)
anyways, it shouldn't be too big a deal. However:
a) There is always the concern of round-off errors occurring when
   doing math/other operations, which could substantially reduce
   accuracy
b) Once we start worrying about vision challenges/buoys more, then
   sub-meter precision (relative to the boat) actually matters.

This does break the reading of old logs (see control/actuator_cmd.proto).

+++ sim/sim_inter.cc
@@ -128,7 +128,7 @@ void SimulatorNode::Iterate() {
 
   Eigen::Vector3d rollpitchyaw = util::GetRollPitchYaw(impl_->get_RBI());
 
-  msg::Vector3f *pos = state_msg_->mutable_pos();
+  msg::Vector3d *pos = state_msg_->mutable_pos();
   pos->set_x(x(0));
   pos->set_y(x(1));
   pos->set_z(x(2));
